# api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: SavewebHQ Project Queue API
  version: 1.0.0
  summary: Single-site PostgreSQL job coordination
...
security:
  - MachineToken: []
paths:
...
  /api/v1/admin/projects:
    get:
      operationId: listAdminProjects
      description: Requires an active user with the admin role.
      responses:
        '200':
          description: Projects and queue counts
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProjectListResponse'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/users:
    get:
      operationId: listAdminUsers
      responses:
        '200':
          description: Users and token state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserListResponse'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/users/{user_id}:
    parameters: [{$ref: '#/components/parameters/UserID'}]
    put:
      operationId: putAdminUser
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminUserRequest'}
      responses:
        '204': {description: User created or updated}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteAdminUser
      description: Deletes the user and cascades to its machine token and web sessions.
      responses:
        '204': {description: User deleted}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/users/{user_id}/machine-token:
    parameters: [{$ref: '#/components/parameters/UserID'}]
    get:
      operationId: getAdminMachineToken
      responses:
        '200':
          description: Current machine token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMachineTokenResponse'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: rotateAdminMachineToken
      responses:
        '201':
          description: New token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminMachineTokenResponse'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: revokeAdminMachineToken
      responses:
        '204': {description: Token revoked}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/projects/{project_id}:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    get:
      operationId: getAdminProject
      description: Requires an active user with the admin role.
      responses:
        '200':
          description: Project and queue counts
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProjectSummary'}
        default: {$ref: '#/components/responses/Error'}
    put:
      operationId: putAdminProject
      description: Creates a project or updates its lifecycle status.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminProjectRequest'}
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProjectSummary'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteAdminProject
      description: Deletes a project and its non-WIP jobs; rejected while WIP jobs exist.
      responses:
        '204': {description: Project deleted}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/projects/{project_id}/jobs:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    get:
      operationId: listAdminProjectJobs
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [todo, wip, done, failed, reset_exhausted]}}
        - {name: after_job_id, in: query, schema: {$ref: '#/components/schemas/JobID'}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        '200':
          description: Project jobs
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminJobListResponse'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: enqueueAdminProjectJobs
      description: Inserts a bounded batch into a non-archived project using its identity mode.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminEnqueueJobsRequest'}
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminEnqueueJobsResponse'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/projects/{project_id}/source:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      operationId: enqueueAdminProjectSource
      description: Streams a jobs-jsonl-zstd-v1 source into a non-archived project using its identity mode.
      requestBody:
        required: true
        content:
          application/zstd:
            schema: {type: string, format: binary}
      responses:
        '200':
          description: Source import result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminEnqueueSourceResponse'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/projects/{project_id}/jobs/{job_id}:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/JobID'}
    get:
      operationId: getAdminProjectJob
      responses:
        '200':
          description: Complete job state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminJob'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteAdminProjectJob
      description: Deletes a job unless it is WIP.
      responses:
        '204': {description: Job deleted}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/admin/projects/{project_id}/jobs/{job_id}/requeue:
    parameters:
      - {$ref: '#/components/parameters/ProjectID'}
      - {$ref: '#/components/parameters/JobID'}
    post:
      operationId: requeueAdminProjectJob
      description: Resets a failed or reset-exhausted job to todo.
      responses:
        '204': {description: Job requeued}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/projects/{project_id}/jobs/claim:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      operationId: claimProjectJobs
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectClaimRequest'}
      responses:
        '200':
          description: Claimed jobs
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectClaimResponse'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/projects/{project_id}/jobs/complete:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      operationId: completeProjectJobs
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectCompleteRequest'}
      responses:
        '200': {$ref: '#/components/responses/BatchResult'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/projects/{project_id}/jobs/fail:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      operationId: failProjectJobs
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectFailRequest'}
      responses:
        '200': {$ref: '#/components/responses/BatchResult'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/projects/{project_id}/jobs/extend-lease:
    parameters: [{$ref: '#/components/parameters/ProjectID'}]
    post:
      operationId: extendProjectJobLeases
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectExtendLeaseRequest'}
      responses:
        '200': {$ref: '#/components/responses/BatchResult'}
        default: {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    MachineToken: {type: http, scheme: bearer}
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
    UserID:
      name: user_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
    JobID:
      name: job_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/JobID'}
  responses:
    BatchResult:
      description: Per-item mutation results
      content:
        application/json:
          schema: {$ref: '#/components/schemas/BatchResultResponse'}
    Error:
      description: Request error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
    Identifier: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$'}
    JobID: {type: integer, format: int64, minimum: 1}
    Attrs: {type: object, additionalProperties: true}
    JobCounts:
      type: object
      additionalProperties: false
      required: [todo, wip, done, failed, reset_exhausted]
      properties:
        todo: {type: integer, format: int64, minimum: 0}
        wip: {type: integer, format: int64, minimum: 0}
        done: {type: integer, format: int64, minimum: 0}
        failed: {type: integer, format: int64, minimum: 0}
        reset_exhausted: {type: integer, format: int64, minimum: 0}
    AdminUserRequest:
      type: object
      additionalProperties: false
      required: [status, roles]
      properties:
        status: {type: string, enum: [pending, active, suspended]}
        roles: {type: array, uniqueItems: true, items: {type: string, enum: [admin, worker]}}
    AdminUserSummary:
      type: object
      additionalProperties: false
      required: [id, status, roles, machine_token_active, machine_token_viewable, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        github_login: {type: string}
        status: {type: string, enum: [pending, active, suspended]}
        roles: {type: array, items: {type: string, enum: [admin, worker]}}
        machine_token_active: {type: boolean}
        machine_token_viewable: {type: boolean}
        created_at: {type: integer, format: int64}
        updated_at: {type: integer, format: int64}
    AdminUserListResponse:
      type: object
      additionalProperties: false
      required: [users]
      properties:
        users: {type: array, items: {$ref: '#/components/schemas/AdminUserSummary'}}
    AdminMachineTokenResponse:
      type: object
      additionalProperties: false
      required: [user_id, token]
      properties:
        user_id: {$ref: '#/components/schemas/Identifier'}
        token: {type: string, minLength: 1}
    AdminProjectRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: {type: string, enum: [active, draining, archived]}
        identity_mode: {type: string, enum: [none, external_id, unique_value]}
        claim_order: {type: string, enum: [fifo, random]}
    AdminProjectSummary:
      type: object
      additionalProperties: false
      required: [id, status, identity_mode, claim_order, job_counts, created_at, updated_at]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        status: {type: string, enum: [active, draining, archived]}
        identity_mode: {type: string, enum: [none, external_id, unique_value]}
        claim_order: {type: string, enum: [fifo, random]}
        job_counts: {$ref: '#/components/schemas/JobCounts'}
        created_at: {type: integer, format: int64}
        updated_at: {type: integer, format: int64}
    AdminProjectListResponse:
      type: object
      additionalProperties: false
      required: [projects]
      properties:
        projects: {type: array, items: {$ref: '#/components/schemas/AdminProjectSummary'}}
    AdminEnqueueJobsRequest:
      type: object
      additionalProperties: false
      required: [jobs]
      properties:
        jobs: {type: array, minItems: 1, items: {$ref: '#/components/schemas/AdminEnqueueJob'}}
    AdminEnqueueJob:
      type: object
      additionalProperties: false
      required: [value]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        value: {type: string, minLength: 1, maxLength: 8192}
        type: {type: string, enum: [seed, asset], default: seed}
        via: {type: [string, 'null'], maxLength: 8192}
        hops: {type: integer, minimum: 0}
        attr: {$ref: '#/components/schemas/Attrs'}
        random_key: {type: integer, format: int32}
    AdminEnqueueJobsResponse:
      type: object
      additionalProperties: false
      required: [project_id, submitted, inserted]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        submitted: {type: integer, minimum: 1}
        inserted: {type: integer, format: int64, minimum: 0}
    AdminEnqueueSourceResponse:
      type: object
      additionalProperties: false
      required: [project_id, jobs, inserted, uncompressed_bytes]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        jobs: {type: integer, format: int64, minimum: 0}
        inserted: {type: integer, format: int64, minimum: 0}
        uncompressed_bytes: {type: integer, format: int64, minimum: 0}
    AdminJob:
      type: object
      additionalProperties: false
      required: [value, job_id, random_key, status, attempt_id, worker_id, lease_expires_at, reset_count, outcome, warc_receipts, execution_error, created_at, updated_at, completed_at]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        value: {type: string, minLength: 1, maxLength: 8192}
        type: {type: string, enum: [seed, asset], default: seed}
        via: {type: [string, 'null'], maxLength: 8192}
        hops: {type: integer, minimum: 0}
        attr: {$ref: '#/components/schemas/Attrs'}
        job_id: {$ref: '#/components/schemas/JobID'}
        random_key: {type: integer, format: int32}
        status: {type: string, enum: [todo, wip, done, failed, reset_exhausted]}
        attempt_id: {oneOf: [{$ref: '#/components/schemas/Identifier'}, {type: 'null'}]}
        worker_id: {oneOf: [{$ref: '#/components/schemas/Identifier'}, {type: 'null'}]}
        lease_expires_at: {type: [integer, 'null'], format: int64}
        reset_count: {type: integer, minimum: 0}
        outcome: {oneOf: [{$ref: '#/components/schemas/Outcome'}, {type: 'null'}]}
        warc_receipts: {type: array, items: {$ref: '#/components/schemas/WARCReceipt'}}
        execution_error: {oneOf: [{$ref: '#/components/schemas/ExecutionError'}, {type: 'null'}]}
        created_at: {type: integer, format: int64}
        updated_at: {type: integer, format: int64}
        completed_at: {type: [integer, 'null'], format: int64}
    AdminJobListResponse:
      type: object
      additionalProperties: false
      required: [jobs, next_after_job_id]
      properties:
        jobs: {type: array, items: {$ref: '#/components/schemas/AdminJob'}}
        next_after_job_id: {type: [integer, 'null'], format: int64, minimum: 1}
...
    ProjectClaimRequest:
      type: object
      additionalProperties: false
      required: [worker_id, max_jobs, lease_seconds, accept_types]
      properties:
        worker_id: {$ref: '#/components/schemas/Identifier'}
        max_jobs: {type: integer, minimum: 1, maximum: 256}
        lease_seconds: {type: integer, format: int64, minimum: 1, maximum: 3600}
        accept_types: {type: array, maxItems: 16, items: {type: string, enum: [seed, asset]}}
    ProjectClaimResponse:
      type: object
      additionalProperties: false
      required: [project_id, jobs, retry_after_ms]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        jobs: {type: array, items: {$ref: '#/components/schemas/ClaimedJob'}}
        retry_after_ms: {type: integer, format: int64, minimum: 0}
    Outcome:
      type: object
      additionalProperties: false
      required: [kind, code, uri, meta]
      properties:
        kind: {type: string, enum: [success, http_error, skipped]}
        code: {type: [integer, 'null'], minimum: 0, maximum: 999}
        uri: {type: [string, 'null'], maxLength: 4096}
        meta: {$ref: '#/components/schemas/Attrs'}
    WARCReceipt:
      type: object
      additionalProperties: false
      required: [id, issuer, object_id, sha256, size_bytes, accepted_at, signature]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        issuer: {type: string, minLength: 1, maxLength: 512}
        object_id: {type: string, minLength: 1, maxLength: 1024}
        sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        size_bytes: {type: integer, format: int64, minimum: 1}
        accepted_at: {type: integer, format: int64, minimum: 1}
        signature: {type: string, minLength: 1, maxLength: 4096}
...
    ProjectCompleteRequest:
      type: object
      additionalProperties: false
      required: [worker_id, items]
      properties:
        worker_id: {$ref: '#/components/schemas/Identifier'}
        items: {type: array, minItems: 1, maxItems: 256, items: {$ref: '#/components/schemas/ProjectCompleteItem'}}
    ExecutionError:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code: {$ref: '#/components/schemas/Identifier'}
        message: {type: string, maxLength: 2048}
        details: {$ref: '#/components/schemas/Attrs'}
...
    ProjectFailRequest:
      type: object
      additionalProperties: false
      required: [worker_id, items]
      properties:
        worker_id: {$ref: '#/components/schemas/Identifier'}
        items: {type: array, minItems: 1, maxItems: 256, items: {$ref: '#/components/schemas/FailItem'}}
...
    ProjectExtendLeaseRequest:
      type: object
      additionalProperties: false
      required: [worker_id, extend_seconds, items]
      properties:
        worker_id: {$ref: '#/components/schemas/Identifier'}
        extend_seconds: {type: integer, format: int64, minimum: 1, maximum: 3600}
        items: {type: array, minItems: 1, maxItems: 256, items: {$ref: '#/components/schemas/AttemptRef'}}
...
    BatchResultResponse:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results: {type: array, items: {$ref: '#/components/schemas/ItemResult'}}
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/APIError'}